Enhanced `/agencies` `GET` methods and tests to include linked data sources
LIBTYFI:
* Reorganized and refactored some schemas
* Refactored some SQLAlchemy models
* Refactored `test_select_subquery_relation`
* Refactored some subquery relation logic
