+++ middleware/schema_and_dto_logic/primary_resource_schemas/data_requests_schemas.py
@@ -16,9 +16,7 @@ from middleware.schema_and_dto_logic.primary_resource_dtos.data_requests_dtos im
     DataRequestLocationInfoPostDTO,
     DataRequestsPutDTO,
 )
-from middleware.schema_and_dto_logic.primary_resource_schemas.data_sources_schemas import (
-    DataSourceExpandedSchema,
-)
+from middleware.schema_and_dto_logic.primary_resource_schemas.data_sources_base_schemas import DataSourceExpandedSchema
 from middleware.schema_and_dto_logic.schema_helpers import (
     create_get_many_schema,
     create_get_by_id_schema,
