+++ tests/test_database_client.py
@@ -321,8 +321,8 @@ def test_select_from_relation_subquery(
         )
     )
 
-    where_mappings = [WhereMapping(column="id", value=data_source_info.id)]
-    subquery_parameters = [
+    where_mappings_for_data_sources = [WhereMapping(column="id", value=data_source_info.id)]
+    subquery_parameters_for_data_sources = [
         SubqueryParameters(
             relation_name=Relations.AGENCIES_EXPANDED.value,
             columns=["id", "submitted_name"],
@@ -330,22 +330,56 @@ def test_select_from_relation_subquery(
         )
     ]
 
+
+    # Test can get agency from data sources, where the `agencies` property is defined
     results = tdc.db_client._select_from_relation(
-        relation_name="data_sources",
+        relation_name=Relations.DATA_SOURCES_EXPANDED.value,
         columns=["id", "name"],
-        where_mappings=where_mappings,
-        subquery_parameters=subquery_parameters,
+        where_mappings=where_mappings_for_data_sources,
+        subquery_parameters=subquery_parameters_for_data_sources,
     )
 
     assert results == [
         {
             "name": data_source_info.name,
             "id": data_source_info.id,
-            "agency_ids": [agency_info.id],
+            # "agency_ids": [agency_info.id],
             "agencies": [
                 {
                     "submitted_name": agency_info.submitted_name,
                     "id": agency_info.id,
+                }
+            ],
+        }
+    ]
+
+    # Test can also get data sources from agency
+    # Which does not have a `data_sources` property
+    # but which is defined via backref in `DataSourceExpanded.agencies
+    where_mappings_for_agencies = [WhereMapping(column="id", value=agency_info.id)]
+    subquery_parameters_for_agencies = [
+        SubqueryParameters(
+            relation_name=Relations.DATA_SOURCES_EXPANDED.value,
+            columns=["id", "name"],
+            linking_column="data_sources",
+        )
+    ]
+    results = tdc.db_client._select_from_relation(
+        relation_name=Relations.AGENCIES_EXPANDED.value,
+        columns=["id", "submitted_name"],
+        where_mappings=where_mappings_for_agencies,
+        subquery_parameters=subquery_parameters_for_agencies,
+    )
+
+    assert results == [
+        {
+            "submitted_name": agency_info.submitted_name,
+            "id": agency_info.id,
+            # "data_source_ids": [data_source_info.id],
+            "data_sources": [
+                {
+                    "name": data_source_info.name,
+                    "id": data_source_info.id,
                 }
             ],
         }
